jekyll pulls in safe_yaml, which breaks guard-minitest. Workaround.
guard-minitest runs the tests in a subprocess and passes the notifier
configuration as a YAML dump in the GUARD_NOTIFIERS environment
variable. This fails if safe_yaml is being used in the child process
as the symbols will be converted to strings. Our workaround is to
load safe_yaml as early as possible through the test_helper.rb file
and enable symbol parsing.